Make the library path of libpq available in ruby
... and add it to the search paths on Windows which lacks rpath.

In ELF based systems the libpq path is stored as rpath in pg_ext.so, so that libpq is found, even if it's not in the library search paths.

On Windows however the rpath mechanism doesn't work and the search path for libpq must be added manually.
This is done per RubyInstaller::Runtime.add_dll_directory.

The libpq path is available in extconf.rb while build only (while gem install), so that we have to pass the string to lib/pg.rb where pg_ext.so is loaded (while require 'pg').
This is a bit tricky, since pg_ext.so is linked to libpq, so that the library search path has to be set first, prior to loading pg_ext.so.
Therefore a second file is needed bypassing pg_ext.so.
That is postgresql_lib_path.rb and it is passed by mkmf's $INSTALLFILES mechanism.

There's an issue with rake-compiler, which (in contrast to gem install) doesn't copy postgresql_lib_path.rb.
This has to be addressed separatelly.

Another approach would be to add a second extension file, but that is more complicated.

Fixes #365