Call PG::Connection#initialize
Previously PG::Connection.new and PG::Connection.connect_start did a direct call to pgconn_s_allocate() bypassing #initialize.
Now this method is called in both cases.

It might be somewhat unexpected that the arguments given to PG::Connection.new() are not directly passed to #initialize.
They are prepared instead by conninfo_parse and DNS resolution and are passed as a Hash to #initialize.
This is by design.

Fixes #478